requirements.txt
aiogram==2.10.1
aiohttp==3.6.2
aioredis==1.3.1
aiosignal==1.2.0
async-timeout==3.0.1
attrs==21.4.0
Babel==2.9.1
bezier==2021.2.12
cap==0.0.114
certifi==2021.10.8
chardet==3.0.4
charset-normalizer==2.0.12
colorama==0.4.4
cycler==0.11.0
enum34==1.1.10
fonttools==4.31.2
frozenlist==1.3.0
hiredis==2.0.0
idna==3.3
key==0.4
kiwisolver==1.4.2
matplotlib==3.5.1
multidict==4.7.6
nicer==0.0.36
numpy==1.22.3
packaging==21.3
Pillow==9.1.0
pkt==0.0.21
pygifsicle==1.0.5
pyparsing==3.0.7
pysqlcipher3==1.0.4
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
scipy==1.8.0
screeninfo==0.8
six==1.16.0
typing==3.7.4.3
x5==1.0.0
x7==0.2.0
x7-geom==0.7.3
x7-lib==0.5.0
x7-testing==0.5.0
x7-view==0.5.0
yarl==1.7.2
